Add random square coloring feature
Added in a random color mode button to the web
page so that when it's clicked it will allow users
to have randomly colored squares on the grid
instead of the default black. Reworked how the
changeColor function works so that a tile that's
been hovered over once before will not change
color again.